+++ example/android/app/src/main/kotlin/com/example/hng_firebase_auth_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.hng_firebase_auth_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()